% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeQ.R
\name{analyzeQ.CT}
\alias{analyzeQ.CT}
\title{Calculate deltaCT values from imported QuantStudio data}
\usage{
analyzeQ.CT(readqtable, housekeepinggene)
}
\arguments{
\item{readqtable}{Imported data table from \code{readQ()} and associated functions}

\item{housekeepinggene}{character string containing housekeeping gene name}
}
\value{
A table of CT measurements, mean, and SD of housekeeping genes and target genes for each samples, as well as the deltaCT between target and housekeeping genes
}
\description{
This function calculates mean CT values for housekeeping and target genes for each sample
and deltaCT values
}
