% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeQ.R
\name{analyzeQ.ddCT}
\alias{analyzeQ.ddCT}
\title{Calculate delta-deltaCT values from deltaCT values}
\usage{
analyzeQ.ddCT(target.table, control.samples = c("Control_1", "Control_2", "Control_3"))
}
\arguments{
\item{target.table}{Processed data table from \code{analyzeQ.CT()}}

\item{control.samples}{vector of control sample names}
}
\value{
A table of calculateed delta-deltaCT values and fold changes appended to the \code{target.table}
}
\description{
This function calculates delta-delta CT values and fold changes for each sample
relative to the mean of a set of control samples
}
