% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeQ.R
\name{analyzeQ}
\alias{analyzeQ}
\title{Calculate delta-deltaCT values from imported QuantStudio data}
\usage{
analyzeQ(target.table, housekeepinggene, control.samples = c("Control_1", "Control_2", "Control_3"))
}
\arguments{
\item{housekeepinggene}{character string containing housekeeping gene name}

\item{control.samples}{vector of control sample names}

\item{target.table}{Processed data table from \code{analyzeQ.CT()}}
}
\value{
A table of CT values, deltaCT, calculated delta-deltaCT values and fold changes for each gene for each sample
}
\description{
This function calculates mean CT values for housekeeping and target genes for each sample
and deltaCT values from imported QuantStudio data, then calculates delta-delta CT values and fold changes
for each sample relative to the mean of a set of control samples
}
