% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQ.R
\name{importCheck}
\alias{importCheck}
\title{Crosscheck Sample Names and Gene Names in the Uploaded Data Table}
\usage{
importCheck(datatable)
}
\arguments{
\item{datatable}{Imported data table from \code{readQ()} and associated functions}
}
\value{
A table of sample names, gene names, and number of measurements for each
}
\description{
This function produces a table of Sample Names and Gene Names from the uploaded
data table with a count of samples for each.
}
