% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQ.R
\name{readQ.file}
\alias{readQ.file}
\title{Import a single QuantStudio output file}
\usage{
readQ.file(x = "/path/to/QuantStudio.xls")
}
\arguments{
\item{x}{A path to a QuantStudio output file (i.e. \code{"/path/to/Quantstudio.xls"})}
}
\value{
A table of raw qPCR data from a QuantStudio output file
}
\description{
This function imports raw qPCR data from a QuantStudio output file
}
