% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQ.R
\name{readQ.list}
\alias{readQ.list}
\title{Import QuantStudio output files from a list of files}
\usage{
readQ.list(x = list("/path/to/QuantStudio1.xls", "/path/to/QuantStudio2.xls", "/path/to/QuantStudio3.xls"))
}
\arguments{
\item{x}{A list of paths to a QuantStudio output file}
}
\value{
A table of raw qPCR data from all imported QuantStudio output files
}
\description{
This function imports raw qPCR data from QuantStudio output files specified in a list
}
